grafana-dashboards/hack/requirements.txt
PyYAML==5.1.2
requests==2.22.0
jsonpath-ng==1.5.2
